[DTD-381] - Applied the spec updates

+++ app/uk/gov/hmrc/timetopayproxy/models/Customer.scala
@@ -22,7 +22,7 @@ import java.time.LocalDate
 
 final case class Customer(quoteType: QuoteType,
                           instalmentStartDate: LocalDate,
-                          instalmentAmount: Int,
+                          instalmentAmount: BigDecimal,
                           frequency: Frequency,
                           duration: Duration,
                           initialPaymentAmount: Int,
